TG-500: Moved pipelineForm to a separate file
